refcator: handler to service-controller - modelvisibility and verify view belongs to the project
